requirements.txt
# Core AI and Automation
groq>=0.4.0
cohere>=4.0.0
python-dotenv>=1.0.0
rich>=13.0.0

# Web Automation and Browser Control
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Audio and Speech
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
soundfile>=0.12.0
noisereduce>=3.0.0

# GUI Framework
PyQt5>=5.15.0

# Utilities
keyboard>=0.13.5
asyncio-mqtt>=0.13.0
stranslate>=1.0.0

# Optional: For advanced features
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0